ci: Improve CI test job speed (#2123 by @joshuayoes)
